Break types into separate files
