+++ Sources/LLVMSwift/VoidType.swift
@@ -1,0 +1,12 @@
+import cllvm
+
+/// The `Void` type represents any value and has no size.
+public struct VoidType: IRType {
+  /// Creates an instance of the `Void` type.
+  public init() {}
+
+  /// Retrieves the underlying LLVM type object.
+  public func asLLVM() -> LLVMTypeRef {
+    return LLVMVoidType()
+  }
+}
